requirements.txt
Django==5.0.6
stripe==12.4.0
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-django==4.9.0